refactor: Add handler as other type
